backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
SQLAlchemy==1.4.54
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
